+++ lib/listview_utils.dart
@@ -1,7 +1,9 @@
 library listview_utils;
 
-export 'src/list_loading.dart';
-export 'src/custom_list_view.dart';
 export 'src/adapters/list_adapter.dart';
 export 'src/adapters/network_list_adapter.dart';
 export 'src/adapters/static_list_adapter.dart';
+export 'src/controller.dart';
+export 'src/custom_list_view.dart';
+export 'src/list_loading.dart';
+export 'src/pagination_listener.dart';