+++ lib/src/custom_list_view.dart
@@ -4,7 +4,9 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
+import '../listview_utils.dart';
 import 'adapters/list_adapter.dart';
+import 'controller.dart';
 import 'types.dart';
 
 enum PaginationMode {
@@ -47,6 +49,8 @@ class CustomListView<T> extends StatefulWidget {
     this.debounceDuration = const Duration(milliseconds: 500),
     double? itemExtend,
     this.distanceToLoadMore = 200,
+    this.listViewController,
+    this.scrollController,
     this.controller,
   })  : assert(adapter != null || itemCount != null),
         this.itemExtend = separatorBuilder == null ? itemExtend : null,
@@ -90,7 +94,7 @@ class CustomListView<T> extends StatefulWidget {
   /// zero and re-load entries from [adapter]
   final AsyncCallback? onRefresh;
 
-  /// Set true if you would like to disable pull to refres
+  /// Set true if you would like to disable pull to refresh
   /// gesture
   final bool disableRefresh;
 
@@ -119,8 +123,15 @@ class CustomListView<T> extends StatefulWidget {
   /// Scroll distance to the end in pixels required to load more
   final double distanceToLoadMore;
 
+  /// The list view controller
+  final CustomListViewController listViewController;
+
   /// Scroll controller
-  final ScrollController? controller;
+  final ScrollController? scrollController;
+
+  @Deprecated(
+      'The controller property has been deprecated. Use scrollController instead.')
+  final ScrollController controller;
 
   @override
   CustomListViewState createState() => CustomListViewState();
@@ -163,6 +174,10 @@ class CustomListViewState extends State<CustomListView> {
 
     if (!loadMore(offset: widget.initialOffset)) {
       _stateNotifier.value = _CLVState.idle;
+    }
+
+    if (widget.listViewController != null) {
+      widget.listViewController.attach(this);
     }
   }
 
@@ -218,6 +233,30 @@ class CustomListViewState extends State<CustomListView> {
     }
   }
 
+  /// Like [refresh], but shows the loading indicator and the error state
+  Future refreshWithIndicator() async {
+    assert(widget.onRefresh != null || widget.adapter != null);
+
+    if (_loadDebounce?.isActive ?? false) _loadDebounce.cancel();
+
+    setState(() => items.clear());
+    _loading = true;
+    _stateNotifier.value = _CLVState.loading;
+
+    try {
+      if (widget.onRefresh != null) {
+        await widget.onRefresh();
+      } else if (widget.adapter != null) {
+        await reload();
+      }
+      _stateNotifier.value = _CLVState.idle;
+    } catch (e) {
+      _stateNotifier.value = _CLVState.createError(e);
+    } finally {
+      _loading = false;
+    }
+  }
+
   /// Returns true if loading was triggered, or false if loading is not
   /// triggered because of might be already loading or there is not a source for
   /// loading data from.
@@ -240,15 +279,6 @@ class CustomListViewState extends State<CustomListView> {
     });
 
     return true;
-  }
-
-  bool handleScrollNotification(ScrollNotification notification) {
-    if (notification.depth <= 0 &&
-        notification.metrics.maxScrollExtent - notification.metrics.pixels <
-            widget.distanceToLoadMore) {
-      loadMore();
-    }
-    return false;
   }
 
   /// Returns item count
@@ -291,7 +321,11 @@ class CustomListViewState extends State<CustomListView> {
           }
         }
 
-        return widget.itemBuilder(context, index, items[index]);
+        return widget.itemBuilder(
+          context,
+          index,
+          widget.adapter != null ? items[index] : null,
+        );
       },
       semanticIndexCallback: (_, int index) {
         if (widget.separatorBuilder != null) {
@@ -344,7 +378,7 @@ class CustomListViewState extends State<CustomListView> {
       scrollDirection: widget.scrollDirection,
       shrinkWrap: widget.shrinkWrap,
       semanticChildCount: itemCount,
-      controller: widget.controller,
+      controller: widget.scrollController ?? widget.controller,
       physics: widget.physics,
       slivers: slivers.map((sliver) {
         int index = i++;
@@ -362,6 +396,11 @@ class CustomListViewState extends State<CustomListView> {
   @override
   void didUpdateWidget(CustomListView old) {
     super.didUpdateWidget(old);
+    if (old.listViewController != widget.listViewController &&
+        widget.listViewController != null) {
+      widget.listViewController.attach(this);
+    }
+
     if (old.adapter != widget.adapter) {
       if (old.adapter != null && widget.adapter != null) {
         if (old.adapter!.shouldUpdate(widget.adapter) == false) {
@@ -374,8 +413,9 @@ class CustomListViewState extends State<CustomListView> {
 
   @override
   Widget build(BuildContext context) {
-    final Widget child = NotificationListener<ScrollNotification>(
-      onNotification: handleScrollNotification,
+    final Widget child = PaginationListener(
+      onNextPage: loadMore,
+      distance: widget.distanceToLoadMore,
       child: _buildList(context),
     );